requirements.txt
langchain==0.1.17
langchain_community==0.0.36
langchain_core==0.1.48
langchain_groq==0.1.3
langchain_nomic==0.0.2
openai==1.25.0
python-dotenv==1.0.1
streamlit==1.31.0
